+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### 🎉 Added
 
 - `@qiskit/qiskit-sim`: Introduce Gate class
+- `@qiskit/qiskit-sim`: Add the identity (id) gate
 
 ### ✏️ Changed